requirements.txt
PyYAML
traits
zcommons=0.1.0
